Update 01 第一章-因果关系入门.ipynb
- Fixed some translation errors in the first half of Chapter 1.
- Transformed some expressions and format to make the chapter read more smoothly.
- Revised the first paragraph in the "Bias" section.
- Adjust some language usage and add some explanation to make sure the whole section is easy to understand.
- Standardized the usage of multiple exchangeable translations for "treated/untreated", and "control group".

* "treatment" -> "处理/未处理 "
* "control group ->"对照组"

I avoided "治疗" for the word "treated" as much as possible, due to the expressions kind of distorted the original meaning of "treatment" under the context of causal inference. I considered using "干预" for "treatment",  especially since the author did not use many "interventions" in the book.  However, it still slightly shifted away from the original meaning of treatment. Besides, "处理效应/处理组" are already widely accepted terms in multiple fields. So I picked "处理" for "treatment", even though it may not be perfect.

- Replaced "单位""单元" with "个体"。
- Further explained "对照组".